fix: the three Windows and coverage failures
**`SCAV_CLIP_NONE` was -1 on Windows.** An unnamed C enum takes an
implementation-chosen underlying type, and under the MS ABI that is `int`,
so a `0xFFFFFFFF` enumerator wraps. clang-cl said so twice -- once as a
narrowing error at the initializer, then as `-Wsign-compare` at every
comparison once the initializer was cast. Casting each site would have left
an ABI constant whose C value differs per platform, which is precisely what
§6 bans typeless enums to prevent, so it is a `#define` now. The other
SCAV_* enums all fit `int` and stay enums.

That exposed a fail-closed hole in the ABI scraper: it skipped every
`#define` in silence, so moving a constant out of an enum would have
dropped it from the golden without a word -- the same class of bug the
scraper was written to catch. It now records object-like macros as
constants, refuses function-like ones, and refuses a value that is not an
integer. Two more cases in the fail-closed test, and the generated bindings
carry SCAV_CLIP_NONE again.

**MSVC's STL needs `<ostream>` to stringify a `string_view`**, which doctest
does for any CHECK on one. libc++ and libstdc++ supply it transitively;
MSVC does not, and the error surfaces inside `__msvc_string_view.hpp` rather
than at the test. Added to the six new test files, matching what the
pre-existing ones already do.

**The coverage gate asked about a generated file.** The embedded font source
lives in the build tree and no test links it, so the gate reported "no
coverage data" for something no one wrote and nothing could cover. The
declared-source manifest now excludes anything under the binary directory.

Verified locally: release, debug, ASan, UBSan, the coverage gate (all 32
declared sources executed), clang-format and clang-tidy clean.

+++ src/draw/drawlist.cpp
@@ -98,9 +98,7 @@ void push_prim(DrawList &d,
   d.prims.push_back({ .kind = kind,
                       .depth = depth,
                       .style = style,
-                      // Cast because a C enum's underlying type is the
-                      // implementation's choice, and MSVC calls this narrowing.
-                      .clip = static_cast<uint32_t>(SCAV_CLIP_NONE),
+                      .clip = SCAV_CLIP_NONE,
                       .origin_kind = static_cast<uint32_t>(origin.kind),
                       .origin_ordinal = origin.ordinal,
                       .points = points,
